% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_kinetics.R
\name{GenerateCycleEVF}
\alias{GenerateCycleEVF}
\title{Generating EVFs for cells sampled along cycle trajectory}
\usage{
GenerateCycleEVF(
  ncells_total,
  nextra,
  n_unstable,
  nevf,
  n_de_evf,
  vary = "all",
  depth = 1,
  evf_center = 1,
  Sigma,
  seed
)
}
\arguments{
\item{ncells_total}{number of cells}

\item{nextra}{number of cells in second cycle}

\item{n_unstable}{number of cells before reaching steady state}

\item{n_de_evf}{Number of EVFs with an impulse signal}

\item{vary}{which parameters are affected by Diff-EVFs. Can be "kon", "koff", "s", "all", "except_kon", "except_koff", "except_s". Suggestions are "all" or "s"}

\item{depth}{depth of the sampling process}

\item{evf_center}{the mean of Gaussain function where the non-Diff EVFs are sampled from}

\item{Sigma}{The standard deviation of the brownian motion of EVFs changing along the tree}

\item{seed}{the random seed}

\item{n_evf}{Number of EVFs}
}
\value{
a list of two object, one is the evf, and the other is a dataframe indicating the branch each cell comes from (pop) and its depth in the tree (depth)
}
\description{
Generating EVFs for cells sampled along cycle trajectory
}
