% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_kinetics.R
\name{GenerateForwardEVF}
\alias{GenerateForwardEVF}
\title{Generating first half cycle of EVFs for cells sampled along cycle trajectory}
\usage{
GenerateForwardEVF(
  ncells,
  n_nd_evf,
  n_de_evf,
  depth = 1,
  evf_center = 1,
  vary = "s",
  Sigma,
  seed
)
}
\arguments{
\item{ncells}{Number of cells}

\item{n_nd_evf}{Number of EVFs that do not have an impulse signal}

\item{n_de_evf}{Number of EVFs with an impulse signal}

\item{depth}{depth of the sampling process}

\item{evf_center}{the mean of Gaussain function where the non-Diff EVFs are sampled from}

\item{vary}{which kinetic parameters should the differential evfs affect. Default is 's'. Can be "kon", "koff", "s", "all", "except_kon", "except_koff", "except_s". Suggestions are "all" or "s".}

\item{Sigma}{The standard deviation of the brownian motion of EVFs changing along the tree}

\item{seed}{the random seed}
}
\value{
a list of two object, one is the evf, and the other is a dataframe indicating the branch each cell comes from (pop) and its depth in the tree (depth)
}
\description{
Generating first half cycle of EVFs for cells sampled along cycle trajectory
}
