% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_counts.R
\name{includeHge}
\alias{includeHge}
\title{Update the kinetic parameters of highly expressed genes}
\usage{
includeHge(prop_hge = 0.015, mean_hge = 5, params)
}
\arguments{
\item{prop_hge}{proportion of the highly expressed gene}

\item{mean_hge}{parameter of highly expressed gene}

\item{params}{kinetic parameters}
}
\value{
updated kinetic parameters
}
\description{
Update the kinetic parameters of highly expressed genes
}
