% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotEVF}
\alias{plotEVF}
\title{Visualize cell evfs, colored using the simulated time}
\usage{
plotEVF(evf_res, width = 15, height = 15, units = "in", dpi = 1000)
}
\arguments{
\item{evf_res}{evf parameters}

\item{width}{width of the figure}

\item{height}{height of the figure}

\item{units}{select from "cm", "mm" and "in", default "in"}

\item{dpi}{dpi value of the figure}

\item{filename}{name of the saved figure}
}
\value{
None
}
\description{
Visualize cell evfs, colored using the simulated time
}
