% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotPseudotime}
\alias{plotPseudotime}
\title{Visualize the result, colored using the simulated time}
\usage{
plotPseudotime(
  filename,
  result,
  dr = "pca",
  width = 15,
  height = 15,
  units = "in",
  dpi = 1000
)
}
\arguments{
\item{filename}{name of the saved figure}

\item{result}{output from the simulator}

\item{dr}{dimension reduction method, default "pca}

\item{width}{width of the figure}

\item{height}{height of the figure}

\item{units}{select from "cm", "mm" and "in", default "in"}

\item{dpi}{dpi value of the figure}
}
\value{
None
}
\description{
Visualize the result, colored using the simulated time
}
