% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotVelo}
\alias{plotVelo}
\title{Visualize the simulated velocity, colored using the simulated time}
\usage{
plotVelo(filename, result, width = 15, height = 15, units = "in", dpi = 1000)
}
\arguments{
\item{filename}{name of the saved figure}

\item{result}{output from the simulator}

\item{width}{width of the figure}

\item{height}{height of the figure}

\item{units}{select from "cm", "mm" and "in", default "in"}

\item{dpi}{dpi value of the figure}

\item{dr}{dimension reduction method, default "pca}
}
\value{
None
}
\description{
Visualize the simulated velocity, colored using the simulated time
}
