% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_counts.R
\name{SimulateVeloCycle}
\alias{SimulateVeloCycle}
\title{Generate both evf and gene effect and simulate true transcript counts, cycle structure}
\usage{
SimulateVeloCycle(
  ncells_total,
  ngenes,
  start_s = NULL,
  start_u = NULL,
  start_state = NULL,
  evf_center = 1,
  nevf = 20,
  randseed,
  n_de_evf = 12,
  vary = "s",
  Sigma = 0.1,
  geffect_mean = 0,
  gene_effects_sd = 1,
  gene_effect_prob = 0.3,
  bimod = 0,
  param_realdata = "zeisel.imputed",
  scale_s = 1,
  prop_hge = 0.015,
  mean_hge = 5,
  nextra = 0,
  n_unstable = 25,
  plot = FALSE
)
}
\arguments{
\item{ncells_total}{number of cells}

\item{ngenes}{number of genes}

\item{start_s}{initial spliced count, NULL if not specified}

\item{start_u}{initial unspliced count, NULL if not specified}

\item{evf_center}{the value which evf mean is generated from}

\item{nevf}{number of evfs}

\item{randseed}{random seed}

\item{n_de_evf}{number of differential evfs between populations}

\item{vary}{which kinetic parameters should the differential evfs affect. Default is 's'. Can be "kon", "koff", "s", "all", "except_kon", "except_koff", "except_s". Suggestions are "all" or "s".}

\item{Sigma}{parameter of the std of evf values within the same population}

\item{geffect_mean}{the mean of the normal distribution where the non-zero gene effect sizes are sampled from}

\item{gene_effect_prob}{the probability that the effect size is not 0}

\item{bimod}{the amount of increased bimodality in the transcript distribution, 0 being not changed from the results calculated using evf and gene effects, and 1 being all genes are bimodal}

\item{param_realdata}{pick from zeisel.imputed or NULL; zeisel.imputed means using the distribution of kinetic parameters learned from the Zeisel 2015 dataset. This option is recommended.}

\item{scale_s}{a factor to scale the s parameter, which is used to tune the size of the actual cell (small cells have less number of transcripts in total)}

\item{prop_hge}{the proportion of very highly expressed genes}

\item{mean_hge}{the parameter to amplify the gene-expression levels of the very highly expressed genes}

\item{n_unstable}{the number of cells to be removed at the beginning of the simulation, sutract from ncells_total}

\item{plot}{plot the kinetic or not}

\item{start_stats}{initial kinetic state, NULL if not specified}

\item{phyla}{the cell developmental tree if chosing 'discrete' or 'continuous' evf type. Can either be generated randomly (using pbtree(nclusters) function from phytools package) or read from newick format file using the ape package}

\item{gene_effect_sd}{the standard deviation of the normal distribution where the non-zero gene effect sizes are sampled from}
}
\value{
a list of 4 elements, the first element is true counts, second is the gene level meta information, the third is cell level meta information, including a matrix of evf and a vector of cell identity, and the fourth is the parameters kon, koff and s used to simulation the true counts
}
\description{
Generate both evf and gene effect and simulate true transcript counts, cycle structure
}
