% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_counts.R
\name{technicalNoise}
\alias{technicalNoise}
\title{Add technical noise to the data}
\usage{
technicalNoise(result, capture.rate = 0.2)
}
\arguments{
\item{result}{result produced by simulation}

\item{capture.rate}{capture rate of RNA molecules, between 0 and 1}
}
\value{
a list of 4 elements, the first element is true counts, second is the gene level meta information, the third is cell level meta information, including a matrix of evf and a vector of cell identity, and the fourth is the parameters kon, koff and s used to simulation the true counts
}
\description{
Add technical noise to the data
}
